requirements.txt
requests>=2.31
python-dotenv>=1.0
tenacity>=8.2
tqdm>=4.66
sqlite3  # stdlib
pybaseball>=2.2
nhl-api-py>=0.9
soccerdata>=1.5
pandas>=2.0
numpy>=1.26
scipy>=1.11
